[NFC] Refactor RunBenchmark()
